+++ pypeit/core/datacube.py
@@ -1106,11 +1106,11 @@ def generate_image_subpixel(image_wcs, all_ra, all_dec, all_wave, all_sci, all_i
     Args:
         image_wcs (`astropy.wcs.wcs.WCS`_):
             World coordinate system to use for the white light images.
-        all_ra (`numpy.ndarray`_)
+        all_ra (`numpy.ndarray`_):
             1D flattened array containing the right ascension of each pixel (units = degrees)
-        all_dec (`numpy.ndarray`_)
+        all_dec (`numpy.ndarray`_):
             1D flattened array containing the declination of each pixel (units = degrees)
-        all_wave (`numpy.ndarray`_)
+        all_wave (`numpy.ndarray`_):
             1D flattened array containing the wavelength of each pixel (units = Angstroms)
         all_sci (`numpy.ndarray`_):
             1D flattened array containing the counts of each pixel from all spec2d files
@@ -1118,22 +1118,22 @@ def generate_image_subpixel(image_wcs, all_ra, all_dec, all_wave, all_sci, all_i
             1D flattened array containing the inverse variance of each pixel from all spec2d files
         all_wghts (`numpy.ndarray`_):
             1D flattened array containing the weights of each pixel to be used in the combination
-        all_spatpos (`numpy.ndarray`_)
+        all_spatpos (`numpy.ndarray`_):
             1D flattened array containing the detector pixel location in the spatial direction
-        all_specpos (`numpy.ndarray`_)
+        all_specpos (`numpy.ndarray`_):
             1D flattened array containing the detector pixel location in the spectral direction
-        all_spatid (`numpy.ndarray`_)
+        all_spatid (`numpy.ndarray`_):
             1D flattened array containing the spatid of each pixel
-        tilts (`numpy.ndarray`_, list)
+        tilts (`numpy.ndarray`_, list):
             2D wavelength tilts frame, or a list of tilt frames (see all_idx)
-        slits (:class:`pypeit.slittrace.SlitTraceSet`_, list)
+        slits (:class:`pypeit.slittrace.SlitTraceSet`, list):
             Information stored about the slits, or a list of SlitTraceSet (see all_idx)
-        astrom_trans (:class:`pypeit.alignframe.AlignmentSplines`_, list):
+        astrom_trans (:class:`pypeit.alignframe.AlignmentSplines`, list):
             A Class containing the transformation between detector pixel coordinates
             and WCS pixel coordinates, or a list of Alignment Splines (see all_idx)
         bins (tuple):
             A 3-tuple (x,y,z) containing the histogram bin edges in x,y spatial and z wavelength coordinates
-        all_idx (`numpy.ndarray`_, optional)
+        all_idx (`numpy.ndarray`_, optional):
             If tilts, slits, and astrom_trans are lists, this should contain a 1D flattened array, of
             the same length as all_sci, containing the index the tilts, slits, and astrom_trans lists
             that corresponds to each pixel. Note that, in this case all of these lists need to be the same length.
@@ -1184,15 +1184,15 @@ def generate_cube_subpixel(outfile, output_wcs, all_ra, all_dec, all_wave, all_s
     for further details about this algorithm
 
     Args:
-        outfile (`str`):
+        outfile (str):
             Filename to be used to save the datacube
         output_wcs (`astropy.wcs.wcs.WCS`_):
             Output world coordinate system.
-        all_ra (`numpy.ndarray`_)
+        all_ra (`numpy.ndarray`_):
             1D flattened array containing the right ascension of each pixel (units = degrees)
-        all_dec (`numpy.ndarray`_)
+        all_dec (`numpy.ndarray`_):
             1D flattened array containing the declination of each pixel (units = degrees)
-        all_wave (`numpy.ndarray`_)
+        all_wave (`numpy.ndarray`_):
             1D flattened array containing the wavelength of each pixel (units = Angstroms)
         all_sci (`numpy.ndarray`_):
             1D flattened array containing the counts of each pixel from all spec2d files
@@ -1200,13 +1200,13 @@ def generate_cube_subpixel(outfile, output_wcs, all_ra, all_dec, all_wave, all_s
             1D flattened array containing the inverse variance of each pixel from all spec2d files
         all_wghts (`numpy.ndarray`_):
             1D flattened array containing the weights of each pixel to be used in the combination
-        all_spatpos (`numpy.ndarray`_)
+        all_spatpos (`numpy.ndarray`_):
             1D flattened array containing the detector pixel location in the spatial direction
-        all_specpos (`numpy.ndarray`_)
+        all_specpos (`numpy.ndarray`_):
             1D flattened array containing the detector pixel location in the spectral direction
-        all_spatid (`numpy.ndarray`_)
+        all_spatid (`numpy.ndarray`_):
             1D flattened array containing the spatid of each pixel
-        tilts (`numpy.ndarray`_, list)
+        tilts (`numpy.ndarray`_, list):
             2D wavelength tilts frame, or a list of tilt frames (see all_idx)
         slits (:class:`pypeit.slittrace.SlitTraceSet`_, list)
             Information stored about the slits, or a list of SlitTraceSet (see all_idx)
@@ -1215,19 +1215,19 @@ def generate_cube_subpixel(outfile, output_wcs, all_ra, all_dec, all_wave, all_s
             and WCS pixel coordinates, or a list of Alignment Splines (see all_idx)
         bins (tuple):
             A 3-tuple (x,y,z) containing the histogram bin edges in x,y spatial and z wavelength coordinates
-        all_idx (`numpy.ndarray`_, optional)
+        all_idx (`numpy.ndarray`_, optional):
             If tilts, slits, and astrom_trans are lists, this should contain a 1D flattened array, of
             the same length as all_sci, containing the index the tilts, slits, and astrom_trans lists
             that corresponds to each pixel. Note that, in this case all of these lists need to be the same length.
-        spec_subpixel (`int`, optional):
+        spec_subpixel (int, optional):
             What is the subpixellation factor in the spectral direction. Higher values give more reliable results,
             but note that the time required goes as (spec_subpixel * spat_subpixel). The default value is 5,
             which divides each detector pixel into 5 subpixels in the spectral direction.
-        spat_subpixel (`int`, optional):
+        spat_subpixel (int, optional):
             What is the subpixellation factor in the spatial direction. Higher values give more reliable results,
             but note that the time required goes as (spec_subpixel * spat_subpixel). The default value is 5,
             which divides each detector pixel into 5 subpixels in the spatial direction.
-        overwrite (`bool`, optional):
+        overwrite (bool, optional):
             If True, the output cube will be overwritten.
         blaze_wave (`numpy.ndarray`_, optional):
             Wavelength array of the spectral blaze function
@@ -1246,7 +1246,7 @@ def generate_cube_subpixel(outfile, output_wcs, all_ra, all_dec, all_wave, all_s
             the minimum/maximum wavelength of all_wave.
         specname (str, optional):
             Name of the spectrograph
-        debug (bool):
+        debug (bool, optional):
             If True, a residuals cube will be output. If the datacube generation is correct, the
             distribution of pixels in the residual cube with no flux should have mean=0 and std=1.
     """